Fix test after enabled metrics on windows
